andres + eric comments
