+++ public/app/plugins/datasource/cloudwatch/components/VariableQueryEditor/VariableQueryField.tsx
@@ -1,9 +1,8 @@
 import { SelectableValue } from '@grafana/data';
-import { InlineField, Input, Select } from '@grafana/ui';
-import React, { FC, useState } from 'react';
+import { InlineField, Select } from '@grafana/ui';
+import React, { FC } from 'react';
 
 const LABEL_WIDTH = 20;
-const TEXT_WIDTH = 100;
 
 interface VariableQueryFieldProps {
   onChange: (value: string) => void;
@@ -35,29 +34,6 @@ export const VariableQueryField: FC<VariableQueryFieldProps> = ({
         options={options}
         isLoading={isLoading}
         inputId="inline-field"
-      />
-    </InlineField>
-  );
-};
-
-interface VariableTextFieldProps {
-  onBlur: (value: string) => void;
-  placeholder: string;
-  value: string;
-  label: string;
-}
-
-export const VariableTextField: FC<VariableTextFieldProps> = ({ label, onBlur, placeholder, value }) => {
-  const [localValue, setLocalValue] = useState(value);
-  return (
-    <InlineField label={label} labelWidth={LABEL_WIDTH}>
-      <Input
-        aria-label={label}
-        placeholder={placeholder}
-        value={localValue}
-        onChange={(e) => setLocalValue(e.currentTarget.value)}
-        onBlur={() => onBlur(localValue)}
-        width={TEXT_WIDTH}
       />
     </InlineField>
   );
