+++ public/app/plugins/datasource/cloudwatch/components/VariableQueryEditor/index.tsx
@@ -1,0 +1,1 @@
+export { VariableQueryEditor } from './VariableQueryEditor';
